requirements.txt
anthropic==0.18.0
pandas==2.1.4
python-dotenv==1.0.0
tqdm==4.66.1 